ft_strlen update/norm

+++ ft_printf/ft_printf_utils.c
@@ -49,7 +49,7 @@ int	ft_print_str(char *str)
 	return (i);
 }
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	int	i;
